cleanup: move gRPC centric code generation to a gRPC specific file
This is a followup to an earlier review where it was stated that
moving gRPC centric code to its own file was a worthwhile maintenance
task.

The diff in this commit is large but is almost entirely moving around
existing code blocks with no structural or behavioral changes.
Any new code is just boilerplate intended to check which transports
the generator should emit code for and dispatch accordingly or
renaming existing methods to indicate their gRPC centric nature.
