+++ internal/gengapic/gengapic.go
@@ -116,20 +116,6 @@ func Gen(genReq *plugin.CodeGeneratorRequest) (*plugin.CodeGeneratorResponse, er
 	return &g.resp, nil
 }
 
-func (g *generator) genGRPCMethods(serv *descriptor.ServiceDescriptorProto, servName string) error {
-	g.addMetadataServiceForTransport(serv.GetName(), "grpc", servName)
-
-	methods := append(serv.GetMethod(), g.getMixinMethods()...)
-	for _, m := range methods {
-		g.methodDoc(m)
-		if err := g.genMethod(servName, serv, m); err != nil {
-			return errors.E(err, "method: %s", m.GetName())
-		}
-		g.addMetadataMethod(serv.GetName(), "grpc", m.GetName())
-	}
-	return nil
-}
-
 // gen generates client for the given service.
 func (g *generator) gen(serv *descriptor.ServiceDescriptorProto) error {
 	servName := pbinfo.ReduceServName(serv.GetName(), g.opts.pkgName)
@@ -207,134 +193,6 @@ type auxTypes struct {
 	// Since multiple methods can page over the same type, we dedupe by the name of the iterator,
 	// which is in turn determined by the element type name.
 	iters map[string]*iterType
-}
-
-// genMethod generates a single method from a client. m must be a method declared in serv.
-// If the generated method requires an auxillary type, it is added to aux.
-func (g *generator) genMethod(servName string, serv *descriptor.ServiceDescriptorProto, m *descriptor.MethodDescriptorProto) error {
-	// Check if the RPC returns google.longrunning.Operation.
-	if g.isLRO(m) {
-		g.aux.lros[m] = true
-		return g.lroCall(servName, m)
-	}
-
-	if m.GetOutputType() == emptyType {
-		return g.emptyUnaryCall(servName, m)
-	}
-
-	if pf, err := g.pagingField(m); err != nil {
-		return err
-	} else if pf != nil {
-		iter, err := g.iterTypeOf(pf)
-		if err != nil {
-			return err
-		}
-
-		return g.pagingCall(servName, m, pf, iter)
-	}
-
-	switch {
-	case m.GetClientStreaming():
-		return g.noRequestStreamCall(servName, serv, m)
-	case m.GetServerStreaming():
-		return g.serverStreamCall(servName, serv, m)
-	default:
-		return g.unaryCall(servName, m)
-	}
-}
-
-func (g *generator) unaryCall(servName string, m *descriptor.MethodDescriptorProto) error {
-	s := g.descInfo.ParentElement[m]
-	sFQN := fmt.Sprintf("%s.%s", g.descInfo.ParentFile[s].GetPackage(), s.GetName())
-	inType := g.descInfo.Type[*m.InputType]
-	outType := g.descInfo.Type[*m.OutputType]
-
-	inSpec, err := g.descInfo.ImportSpec(inType)
-	if err != nil {
-		return err
-	}
-	outSpec, err := g.descInfo.ImportSpec(outType)
-	if err != nil {
-		return err
-	}
-
-	p := g.printf
-
-	lowcaseServName := lowerFirst(servName + "GRPCClient")
-
-	p("func (c *%s) %s(ctx context.Context, req *%s.%s, opts ...gax.CallOption) (*%s.%s, error) {",
-		lowcaseServName, m.GetName(), inSpec.Name, inType.GetName(), outSpec.Name, outType.GetName())
-
-	g.deadline(sFQN, m.GetName())
-
-	err = g.insertMetadata(m)
-	if err != nil {
-		return err
-	}
-	g.appendCallOpts(m)
-
-	p("var resp *%s.%s", outSpec.Name, outType.GetName())
-	p("err := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {")
-	p("  var err error")
-	p("  resp, err = %s", g.grpcStubCall(m))
-	p("  return err")
-	p("}, opts...)")
-	p("if err != nil {")
-	p("  return nil, err")
-	p("}")
-	p("return resp, nil")
-
-	p("}")
-	p("")
-
-	g.imports[inSpec] = true
-	g.imports[outSpec] = true
-
-	return nil
-}
-
-func (g *generator) emptyUnaryCall(servName string, m *descriptor.MethodDescriptorProto) error {
-	s := g.descInfo.ParentElement[m]
-	sFQN := fmt.Sprintf("%s.%s", g.descInfo.ParentFile[s].GetPackage(), s.GetName())
-	inType := g.descInfo.Type[*m.InputType]
-
-	inSpec, err := g.descInfo.ImportSpec(inType)
-	if err != nil {
-		return err
-	}
-
-	p := g.printf
-
-	lowcaseServName := lowerFirst(servName + "GRPCClient")
-
-	p("func (c *%s) %s(ctx context.Context, req *%s.%s, opts ...gax.CallOption) error {",
-		lowcaseServName, m.GetName(), inSpec.Name, inType.GetName())
-
-	g.deadline(sFQN, m.GetName())
-
-	err = g.insertMetadata(m)
-	if err != nil {
-		return err
-	}
-	g.appendCallOpts(m)
-	p("err := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {")
-	p("  var err error")
-	p("  _, err = %s", g.grpcStubCall(m))
-	p("  return err")
-	p("}, opts...)")
-	p("return err")
-
-	p("}")
-	p("")
-
-	g.imports[inSpec] = true
-	return nil
-}
-
-func (g *generator) grpcStubCall(method *descriptor.MethodDescriptorProto) string {
-	service := g.descInfo.ParentElement[method]
-	stub := pbinfo.ReduceServName(service.GetName(), g.opts.pkgName)
-	return fmt.Sprintf("c.%s.%s(ctx, req, settings.GRPC...)", grpcClientField(stub), method.GetName())
 }
 
 func (g *generator) deadline(s, m string) {
